Module 6: parallel vascular/cerebrovascular risk side-pipeline
Second, independent pipeline off Module 1's quality gate + Module 2's
segmentation (own try/catch in run_pipeline.m -- never blocks DR
grading). Adds computeTortuosity.m, computeFractalDimension.m,
classifyHypertensiveRetinopathy.m, computeCerebrovascularRiskScore.m
(renamed from computeStrokeRiskScore.m to reflect its actual scope:
stroke + vascular dementia risk, both sharing the small-vessel-disease
mechanism the biomarkers measure -- explicitly NOT an Alzheimer's
detector, different pathology, needs OCT/amyloid imaging this project
doesn't have), and assessVascularRisk.m as the Module 2 wiring point.

Two real bugs found and fixed via end-to-end testing against the real
sample images, not just static analysis:
- bwdist returned single in this MATLAB session (documented default is
  double); strel() rejected it. Cast explicitly in
  classifyVesselSegments.m.
- computeTortuosity.m's mean aggregation was dominated by a few
  branch-point-removal artifacts (near-zero chord length, real arc
  length) producing >6.0 outlier ratios against a same-image median of
  ~0.4. Fixed with median aggregation + a minimum-chord-length floor;
  post-fix values landed at a plausible 0.08-0.10 on both real samples.

README updated with full Module 6 section including both fixes and the
Alzheimer's-scope caveat.

+++ matlab/modules/module6_vascular_risk/classifyVesselSegments.m
@@ -49,7 +49,11 @@ MIN_SEGMENT_LENGTH = 10; % pixels; shorter fragments are mostly noise from the b
 segments = struct('pixelIdx', {}, 'endpoints', {}, 'chordLength', {}, ...
     'arcLength', {}, 'meanWidth', {}, 'colorScore', {}, 'distFromDisc', {}, 'label', {});
 
-distTransform = bwdist(~vesselMask);
+% bwdist has been observed to return single here (toolbox-version-
+% dependent -- its documented default is double) rather than double;
+% cast explicitly since strel() below requires double and silently
+% erroring on a class mismatch is worse than one extra cast.
+distTransform = double(bwdist(~vesselMask));
 red = double(img(:, :, 1));
 green = double(img(:, :, 2));
